Put shared service consts and types into dedicated files
