+++ internal/services/reply.go
@@ -6,26 +6,10 @@ import (
 	"fmt"
 	"net/http"
 
-	g "github.com/dannyh79/brp-webhook/internal/groups"
 	"github.com/dannyh79/brp-webhook/internal/sentry"
 )
 
 var _ Service[GroupDto] = (*ReplyService)(nil)
-
-const lineReplyApiEndpoint = "https://api.line.me/v2/bot/message/reply"
-
-const msgOk = g.MsgRegistrationOk
-const msgAlreadyRegistered = g.MsgAlreadyRegistered
-
-type ReplyMessageRequest struct {
-	ReplyToken string    `json:"replyToken"`
-	Messages   []message `json:"messages"`
-}
-
-type message struct {
-	Type string `json:"type"`
-	Text string `json:"text"`
-}
 
 type ReplyService struct {
 	token  string
